rename: selectComponent 폴더 이동
